feat: Implement Search Result Content and Data Hook for Discovery Module
- Added SearchResultContent component to display search results, bookmarks, and history in the Discovery module.
- Introduced useSearchModalData hook to manage search data, including fetching bookmarks and history, and processing search results.
- Refactored SearchModal to utilize the new component and hook for improved code organization and readability.
