+++ packages/kit/src/views/Discovery/pages/SearchModal/index.tsx
@@ -1,50 +1,24 @@
-import { useCallback, useEffect, useRef, useState } from 'react';
+import { useRef, useState } from 'react';
 
 import { useFocusEffect, useRoute } from '@react-navigation/core';
 import { useIntl } from 'react-intl';
-import { Keyboard, StyleSheet } from 'react-native';
+import { Keyboard } from 'react-native';
 
-import {
-  Icon,
-  Image,
-  Page,
-  RichSizeableText,
-  ScrollView,
-  SearchBar,
-  SizableText,
-  Skeleton,
-  Stack,
-  XStack,
-  useMedia,
-} from '@onekeyhq/components';
-import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
-import { ListItem } from '@onekeyhq/kit/src/components/ListItem';
+import { Page, ScrollView, SearchBar, Stack } from '@onekeyhq/components';
 import useAppNavigation from '@onekeyhq/kit/src/hooks/useAppNavigation';
-import { usePromiseResult } from '@onekeyhq/kit/src/hooks/usePromiseResult';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 import { EEnterMethod } from '@onekeyhq/shared/src/logger/scopes/discovery/scenes/dapp';
-import type { IDiscoveryModalParamList } from '@onekeyhq/shared/src/routes';
-import {
+import type {
   EDiscoveryModalRoutes,
-  EModalRoutes,
+  IDiscoveryModalParamList,
 } from '@onekeyhq/shared/src/routes';
-import type { IDApp } from '@onekeyhq/shared/types/discovery';
 
-import { DiscoveryIcon } from '../../components/DiscoveryIcon';
+import { SearchResultContent } from '../../components/SearchResultContent';
+import { useSearchModalData } from '../../hooks/useSearchModalData';
 import { useWebSiteHandler } from '../../utils/useWebSiteHandler';
 import { withBrowserProvider } from '../Browser/WithBrowserProvider';
 
-import { DappSearchModalSectionHeader } from './DappSearchModalSectionHeader';
-
 import type { RouteProp } from '@react-navigation/core';
-
-const SEARCH_ITEM_ID = 'SEARCH_ITEM_ID';
-
-const LoadingSkeleton = (
-  <Image.Loading>
-    <Skeleton width="100%" height="100%" />
-  </Image.Loading>
-);
 
 function SearchModal() {
   const intl = useIntl();
@@ -56,31 +30,18 @@ function SearchModal() {
   const { useCurrentWindow, tabId, url = '' } = route.params ?? {};
 
   const [searchValue, setSearchValue] = useState(url);
-  const { gtMd } = useMedia();
   const handleWebSite = useWebSiteHandler();
 
-  const { serviceDiscovery } = backgroundApiProxy;
-  const { result: localData, run: refreshLocalData } =
-    usePromiseResult(async () => {
-      const bookmarkData = await serviceDiscovery.getBookmarkData({
-        generateIcon: true,
-        sliceCount: 6,
-      });
-      const historyData = await serviceDiscovery.getHistoryData({
-        generateIcon: true,
-        sliceCount: 6,
-        keyword: searchValue ?? undefined,
-      });
-      return {
-        bookmarkData,
-        historyData,
-      };
-    }, [serviceDiscovery, searchValue]);
-
-  const { result: searchResult } = usePromiseResult(async () => {
-    const res = await serviceDiscovery.searchDApp(searchValue);
-    return res;
-  }, [searchValue, serviceDiscovery]);
+  // Use the extracted hook for data handling
+  const {
+    localData,
+    refreshLocalData,
+    searchList,
+    displaySearchList,
+    displayBookmarkList,
+    displayHistoryList,
+    SEARCH_ITEM_ID,
+  } = useSearchModalData(searchValue);
 
   const jumpPageRef = useRef(false);
   useFocusEffect(() => {
@@ -91,100 +52,6 @@ function SearchModal() {
       jumpPageRef.current = false;
     }
   });
-
-  const [searchList, setSearchList] = useState<IDApp[]>([]);
-
-  useEffect(() => {
-    void (async () => {
-      if (!searchValue) {
-        setSearchList([]);
-        return;
-      }
-      const logo = 'https://uni.onekey-asset.com/static/logo/google.png';
-      const exactUrlResults =
-        searchResult?.filter((item) => item.isExactUrl) || [];
-      const otherResults =
-        searchResult?.filter((item) => !item.isExactUrl) || [];
-      setSearchList([
-        ...exactUrlResults,
-        {
-          dappId: SEARCH_ITEM_ID,
-          name: `${intl.formatMessage({
-            id: ETranslations.explore_search_placeholder,
-          })} "${searchValue}"`,
-          url: '',
-          logo,
-        } as IDApp,
-        ...otherResults,
-      ]);
-    })();
-  }, [searchValue, searchResult, intl]);
-
-  const displaySearchList = Array.isArray(searchList) && searchList.length > 0;
-  const displayBookmarkList =
-    (localData?.bookmarkData ?? []).length > 0 && !displaySearchList;
-  const displayHistoryList = (localData?.historyData ?? []).length > 0;
-
-  const renderList = useCallback(
-    (list: IDApp[]) =>
-      list.map((item, index) => (
-        <ListItem
-          key={index}
-          avatarProps={{
-            src: item.logo || item.originLogo,
-            loading: LoadingSkeleton,
-            fallbackProps: {
-              bg: '$bgStrong',
-              justifyContent: 'center',
-              alignItems: 'center',
-              children: <Icon name="GlobusOutline" />,
-            },
-            borderWidth: StyleSheet.hairlineWidth,
-            borderColor: '$borderSubdued',
-          }}
-          renderItemText={() => (
-            <RichSizeableText
-              linkList={{ a: { url: undefined, cursor: 'auto' } }}
-              numberOfLines={1}
-              size="$bodyLgMedium"
-              flex={1}
-            >
-              {item?.keyword
-                ? item.name.replace(
-                    new RegExp(item.keyword, 'ig'),
-                    `<a>${item.keyword}</a>`,
-                  )
-                : item.name}
-            </RichSizeableText>
-          )}
-          subtitleProps={{
-            numberOfLines: 1,
-          }}
-          onPress={() => {
-            if (item.dappId === SEARCH_ITEM_ID) {
-              handleWebSite({
-                webSite: {
-                  url: searchValue,
-                  title: searchValue,
-                },
-                useCurrentWindow,
-                tabId,
-                enterMethod: EEnterMethod.search,
-              });
-            } else {
-              handleWebSite({
-                dApp: item,
-                useCurrentWindow,
-                tabId,
-                enterMethod: EEnterMethod.search,
-              });
-            }
-          }}
-          testID={`dapp-search${index}`}
-        />
-      )),
-    [handleWebSite, searchValue, tabId, useCurrentWindow],
-  );
 
   return (
     <Page safeAreaEnabled>
@@ -228,119 +95,18 @@ function SearchModal() {
           keyboardShouldPersistTaps="handled"
           onScrollBeginDrag={Keyboard.dismiss}
         >
-          {displaySearchList ? renderList(searchList) : null}
-
-          {displayBookmarkList ? (
-            <Stack>
-              <DappSearchModalSectionHeader
-                title={intl.formatMessage({
-                  id: ETranslations.explore_bookmarks,
-                })}
-                onMorePress={() => {
-                  jumpPageRef.current = true;
-                  navigation.pushModal(EModalRoutes.DiscoveryModal, {
-                    screen: EDiscoveryModalRoutes.BookmarkListModal,
-                  });
-                }}
-              />
-              <XStack $gtMd={{ px: '$3' }}>
-                {localData?.bookmarkData?.map((item, index) => (
-                  <Stack
-                    key={index}
-                    flexBasis="25%"
-                    alignItems="center"
-                    py="$2"
-                    $gtMd={{
-                      flexBasis: '16.66666667%',
-                    }}
-                    onPress={() => {
-                      handleWebSite({
-                        webSite: {
-                          url: item.url,
-                          title: item.title,
-                        },
-                        useCurrentWindow,
-                        tabId,
-                        enterMethod: EEnterMethod.bookmarkInSearch,
-                      });
-                    }}
-                  >
-                    <DiscoveryIcon
-                      uri={item.logo}
-                      size="$14"
-                      borderRadius="$3"
-                    />
-                    <SizableText
-                      mt="$2"
-                      px="$2"
-                      size="$bodyLgMedium"
-                      textAlign="center"
-                      $gtMd={{
-                        size: '$bodyMdMedium',
-                      }}
-                      numberOfLines={1}
-                    >
-                      {item.title}
-                    </SizableText>
-                  </Stack>
-                ))}
-              </XStack>
-            </Stack>
-          ) : null}
-          {displayHistoryList ? (
-            <Stack pt="$5">
-              <DappSearchModalSectionHeader
-                title={intl.formatMessage({
-                  id: ETranslations.browser_recently_closed,
-                })}
-                onMorePress={() => {
-                  jumpPageRef.current = true;
-                  navigation.pushModal(EModalRoutes.DiscoveryModal, {
-                    screen: EDiscoveryModalRoutes.HistoryListModal,
-                  });
-                }}
-              />
-              {localData?.historyData.map((item, index) => (
-                <ListItem
-                  key={index}
-                  avatarProps={{
-                    src: item.logo,
-                    loading: LoadingSkeleton,
-                    fallbackProps: {
-                      bg: '$bgStrong',
-                      justifyContent: 'center',
-                      alignItems: 'center',
-                      children: <Icon name="GlobusOutline" />,
-                    },
-                    borderWidth: StyleSheet.hairlineWidth,
-                    borderColor: '$borderSubdued',
-                  }}
-                  title={item.title}
-                  titleMatch={item.titleMatch}
-                  titleProps={{
-                    numberOfLines: 1,
-                  }}
-                  subtitle={item.url}
-                  subTitleMatch={item.urlMatch}
-                  subtitleProps={{
-                    numberOfLines: 1,
-                  }}
-                  testID={`search-modal-${item.title.toLowerCase()}`}
-                  onPress={() => {
-                    handleWebSite({
-                      webSite: {
-                        url: item.url,
-                        title: item.title,
-                      },
-                      useCurrentWindow,
-                      tabId,
-                      enterMethod: EEnterMethod.historyInSearch,
-                    });
-                  }}
-                />
-              ))}
-            </Stack>
-          ) : null}
+          <SearchResultContent
+            searchValue={searchValue}
+            localData={localData}
+            searchList={searchList}
+            displaySearchList={displaySearchList}
+            displayBookmarkList={displayBookmarkList}
+            displayHistoryList={displayHistoryList}
+            SEARCH_ITEM_ID={SEARCH_ITEM_ID}
+            handleWebSite={handleWebSite}
+            useCurrentWindow={useCurrentWindow}
+            tabId={tabId}
+          />
         </ScrollView>
       </Page.Body>
     </Page>
